backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
requests>=2.32.0
tenacity>=8.2.0
pybreaker>=0.7.0
weaviate-client>=4.6.0
langgraph>=0.2.0
PyPDF2>=3.0.0
python-docx>=1.1.0
llama-index>=0.10.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
llama-index-embeddings-langchain==0.4.1
llama-index-embeddings-dashscope==0.4.1